+++ CHANGELOG.md
@@ -5,7 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-- Nothing so far
+- Automatically parse all sent messages (e.g. allow `@someone` or `#channel`)
 
 ## [v0.4.0] - 2019-03-18
 ### Changed